move unrelated helpers out of snap_yaml.go and close to where they are used
