+++ snappy/overlord.go
@@ -23,13 +23,16 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"time"
 
 	"github.com/ubuntu-core/snappy/arch"
 	"github.com/ubuntu-core/snappy/dirs"
 	"github.com/ubuntu-core/snappy/logger"
+	"github.com/ubuntu-core/snappy/osutil"
 	"github.com/ubuntu-core/snappy/progress"
 	"github.com/ubuntu-core/snappy/snap"
 	"github.com/ubuntu-core/snappy/snap/squashfs"
+	"github.com/ubuntu-core/snappy/systemd"
 )
 
 // Overlord is responsible for the overall system state.
@@ -104,6 +107,49 @@ func SetupSnap(snapFilePath string, flags InstallFlags, meter progress.Meter) (s
 	}
 
 	return instdir, err
+}
+
+func addSquashfsMount(s *snap.Info, inhibitHooks bool, inter interacter) error {
+	baseDir := s.BaseDir()
+	// XXX: fix BlobPath
+	squashfsPath := stripGlobalRootDir(squashfs.BlobPath(baseDir))
+	whereDir := stripGlobalRootDir(baseDir)
+
+	sysd := systemd.New(dirs.GlobalRootDir, inter)
+	mountUnitName, err := sysd.WriteMountUnitFile(s.Name(), squashfsPath, whereDir)
+	if err != nil {
+		return err
+	}
+
+	// we always enable the mount unit even in inhibit hooks
+	if err := sysd.Enable(mountUnitName); err != nil {
+		return err
+	}
+
+	if !inhibitHooks {
+		return sysd.Start(mountUnitName)
+	}
+
+	return nil
+}
+
+func removeSquashfsMount(baseDir string, inter interacter) error {
+	sysd := systemd.New(dirs.GlobalRootDir, inter)
+	unit := systemd.MountUnitPath(stripGlobalRootDir(baseDir), "mount")
+	if osutil.FileExists(unit) {
+		// we ignore errors, nothing should stop removals
+		if err := sysd.Disable(filepath.Base(unit)); err != nil {
+			logger.Noticef("Failed to disable %q: %s, but continuing anyway.", unit, err)
+		}
+		if err := sysd.Stop(filepath.Base(unit), time.Duration(1*time.Second)); err != nil {
+			logger.Noticef("Failed to stop %q: %s, but continuing anyway.", unit, err)
+		}
+		if err := os.Remove(unit); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
 
 func UndoSetupSnap(installDir string, meter progress.Meter) {
@@ -454,6 +500,39 @@ func canInstall(s *snap.Info, blobf snap.File, allowGadget bool, inter interacte
 
 	if err := checkLicenseAgreement(s, blobf, curr, inter); err != nil {
 		return err
+	}
+
+	return nil
+}
+
+// checkLicenseAgreement returns nil if it's ok to proceed with installing the
+// package, as deduced from the license agreement (which might involve asking
+// the user), or an error that explains the reason why installation should not
+// proceed.
+func checkLicenseAgreement(s *snap.Info, blobf snap.File, cur *snap.Info, ag agreer) error {
+	if s.LicenseAgreement != "explicit" {
+		return nil
+	}
+
+	if ag == nil {
+		return ErrLicenseNotAccepted
+	}
+
+	license, err := blobf.MetaMember("license.txt")
+	if err != nil || len(license) == 0 {
+		return ErrLicenseNotProvided
+	}
+
+	// don't ask for the license if
+	// * the previous version also asked for license confirmation, and
+	// * the license version is the same
+	if cur != nil && (cur.LicenseAgreement == "explicit") && cur.LicenseVersion == s.LicenseVersion {
+		return nil
+	}
+
+	msg := fmt.Sprintf("%s requires that you accept the following license before continuing", s.Name())
+	if !ag.Agreed(msg, string(license)) {
+		return ErrLicenseNotAccepted
 	}
 
 	return nil
